chore: moved akamai plugin tests to akamai folder
