+++ lambda/functions/spf-update/index.js
@@ -13,7 +13,7 @@ function callSpfTools(cb) {
     if (error) {
       // assuming the compare script failed due to the records not matching
       if (error.code == 1) {
-        var update_command = `despf.sh ${process.env.ORIG_SPF} | normalize.sh | simplify.sh | mkblocks.sh ${process.env.DOMAIN} | DOMAIN=${process.env.DOMAIN} ORIG_SPF=${process.env.ORIG_SPF} DRY_RUN=${process.env.DRY_RUN} node ./scripts/plugins/akamai.js`;
+        var update_command = `despf.sh ${process.env.ORIG_SPF} | normalize.sh | simplify.sh | mkblocks.sh ${process.env.DOMAIN} | DOMAIN=${process.env.DOMAIN} ORIG_SPF=${process.env.ORIG_SPF} DRY_RUN=${process.env.DRY_RUN} node ./scripts/plugins/akamai/akamai.js`;
         console.log("Running", update_command);
         child_process.exec(update_command, options, (error, stdout, stderr) => {
           return cb(error, stdout, stderr);
